+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsHeader.base.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { findDOMNode } from 'react-dom';
-import { BaseComponent, css, getRTL, getId, KeyCodes, IRenderFunction, IClassNames } from '../../Utilities';
+import { initializeComponentRef, EventGroup, css, getRTL, getId, KeyCodes, IClassNames } from '../../Utilities';
 import {
   IColumn,
   IDetailsHeaderBaseProps,
@@ -10,7 +10,7 @@ import {
   CheckboxVisibility
 } from './DetailsList.types';
 import { IFocusZone, FocusZone, FocusZoneDirection } from '../../FocusZone';
-import { Icon } from '../../Icon';
+import { Icon, FontIcon } from '../../Icon';
 import { Layer } from '../../Layer';
 import { GroupSpacer } from '../GroupedList/GroupSpacer';
 import { CollapseAllVisibility } from '../../GroupedList';
@@ -30,7 +30,7 @@ const MOUSEMOVE_PRIMARY_BUTTON = 1; // for mouse move event we are using ev.butt
 
 const NO_COLUMNS: IColumn[] = [];
 
-export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, IDetailsHeaderState> implements IDetailsHeader {
+export class DetailsHeaderBase extends React.Component<IDetailsHeaderBaseProps, IDetailsHeaderState> implements IDetailsHeader {
   public static defaultProps = {
     selectAllVisibility: SelectAllVisibility.visible,
     collapseAllVisibility: CollapseAllVisibility.visible
@@ -38,6 +38,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
 
   private _classNames: IClassNames<IDetailsHeaderStyles>;
   private _rootElement: HTMLElement | undefined;
+  private _events: EventGroup;
   private _rootComponent = React.createRef<IFocusZone>();
   private _id: string;
   private _draggedColumnIndex = -1;
@@ -69,6 +70,10 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
 
   constructor(props: IDetailsHeaderBaseProps) {
     super(props);
+
+    initializeComponentRef(this);
+    this._events = new EventGroup(this);
+
     const columnReorderProps: IColumnReorderHeaderProps | undefined =
       props.columnReorderProps || (props.columnReorderOptions && getLegacyColumnReorderProps(props.columnReorderOptions));
     this.state = {
@@ -79,19 +84,6 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
       isAllSelected: !!this.props.selection && this.props.selection.isAllSelected()
     };
 
-    this._onToggleCollapseAll = this._onToggleCollapseAll.bind(this);
-    this._onSelectAllClicked = this._onSelectAllClicked.bind(this);
-    this._updateDragInfo = this._updateDragInfo.bind(this);
-    this._onDragOver = this._onDragOver.bind(this);
-    this._onDrop = this._onDrop.bind(this);
-    this._getHeaderDragDropOptions = this._getHeaderDragDropOptions.bind(this);
-    this._updateDroppingState = this._updateDroppingState.bind(this);
-    this._getDropHintPositions = this._getDropHintPositions.bind(this);
-    this._computeDropHintToBeShown = this._computeDropHintToBeShown.bind(this);
-    this._resetDropHints = this._resetDropHints.bind(this);
-    this._isValidCurrentDropHintIndex = this._isValidCurrentDropHintIndex.bind(this);
-    this._onRootRef = this._onRootRef.bind(this);
-    this._isEventOnHeader = this._isEventOnHeader.bind(this);
     this._onDropIndexInfo = {
       sourceIndex: -1,
       targetIndex: -1
@@ -148,6 +140,8 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     if (this._dragDropHelper) {
       this._dragDropHelper.dispose();
     }
+
+    this._events.dispose();
   }
 
   public render(): JSX.Element {
@@ -165,7 +159,8 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
       onRenderColumnHeaderTooltip = this._onRenderColumnHeaderTooltip,
       styles,
       theme,
-      onRenderDetailsCheckbox
+      onRenderDetailsCheckbox,
+      useFastIcons
     } = this.props;
     const { isAllSelected, columnResizeDetails, isSizing, groupNestingDepth, isAllCollapsed, columnReorderProps } = this.state;
     const showCheckbox = selectAllVisibility !== SelectAllVisibility.none;
@@ -199,6 +194,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     });
 
     const classNames = this._classNames;
+    const IconComponent = useFastIcons ? FontIcon : Icon;
 
     const isRTL = getRTL();
     return (
@@ -249,6 +245,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
                         canSelect={!isCheckboxHidden}
                         className={classNames.check}
                         onRenderDetailsCheckbox={onRenderDetailsCheckbox}
+                        useFastIcons={useFastIcons}
                       />
                     )
                   },
@@ -257,11 +254,11 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
               </div>,
               !this.props.onRenderColumnHeaderTooltip ? (
                 ariaLabelForSelectAllCheckbox && !isCheckboxHidden ? (
-                  <label key="__checkboxLabel" id={`${this._id}-checkTooltip`} className={classNames.accessibleLabel}>
+                  <label key="__checkboxLabel" id={`${this._id}-checkTooltip`} className={classNames.accessibleLabel} aria-hidden={true}>
                     {ariaLabelForSelectAllCheckbox}
                   </label>
                 ) : ariaLabelForSelectionColumn && isCheckboxHidden ? (
-                  <label key="__checkboxLabel" id={`${this._id}-checkTooltip`} className={classNames.accessibleLabel}>
+                  <label key="__checkboxLabel" id={`${this._id}-checkTooltip`} className={classNames.accessibleLabel} aria-hidden={true}>
                     {ariaLabelForSelectionColumn}
                   </label>
                 ) : null
@@ -277,7 +274,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
             aria-expanded={!isAllCollapsed}
             role={ariaLabelForToggleAllGroupsButton ? 'button' : undefined}
           >
-            <Icon className={classNames.collapseButton} iconName={isRTL ? 'ChevronLeftMed' : 'ChevronRightMed'} />
+            <IconComponent className={classNames.collapseButton} iconName={isRTL ? 'ChevronLeftMed' : 'ChevronRightMed'} />
           </div>
         ) : null}
         <GroupSpacer indentWidth={indentWidth} count={groupNestingDepth! - 1} />
@@ -303,6 +300,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
               onRenderColumnHeaderTooltip={this.props.onRenderColumnHeaderTooltip}
               isDropped={this._onDropIndexInfo.targetIndex === columnIndex}
               cellStyleProps={this.props.cellStyleProps}
+              useFastIcons={useFastIcons}
             />,
             this._renderColumnDivider(columnIndex)
           ];
@@ -337,26 +335,26 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     return options;
   }
 
-  private _updateDroppingState(newValue: boolean, event: DragEvent): void {
+  private _updateDroppingState = (newValue: boolean, event: DragEvent): void => {
     if (this._draggedColumnIndex >= 0 && event.type !== 'drop') {
       if (!newValue) {
         this._resetDropHints();
       }
     }
-  }
+  };
 
   private _isValidCurrentDropHintIndex() {
     return this._currentDropHintIndex >= 0;
   }
 
-  private _onDragOver(item: any, event: DragEvent): void {
+  private _onDragOver = (item: any, event: DragEvent): void => {
     if (this._draggedColumnIndex >= 0) {
       event.stopPropagation();
       this._computeDropHintToBeShown(event.clientX);
     }
-  }
+  };
 
-  private _onDrop(item?: any, event?: DragEvent): void {
+  private _onDrop = (item?: any, event?: DragEvent): void => {
     const { columnReorderProps } = this.state;
 
     // Target index will not get changed if draggeditem is after target item.
@@ -384,7 +382,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     this._resetDropHints();
     this._dropHintDetails = {};
     this._draggedColumnIndex = -1;
-  }
+  };
 
   /**
    * @returns whether or not the "Select All" checkbox column is hidden.
@@ -395,7 +393,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     return selectionMode === SelectionMode.none || checkboxVisibility === CheckboxVisibility.hidden;
   }
 
-  private _updateDragInfo(props: { itemIndex: number }, event?: MouseEvent) {
+  private _updateDragInfo = (props: { itemIndex: number }, event?: MouseEvent) => {
     const { columnReorderProps } = this.state;
     const itemIndex = props.itemIndex;
     if (itemIndex >= 0) {
@@ -414,7 +412,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
         columnReorderProps.onColumnDragEnd({ dropLocation: columnDragEndLocation }, event);
       }
     }
-  }
+  };
 
   private _resetDropHints(): void {
     if (this._currentDropHintIndex >= 0) {
@@ -616,9 +614,10 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
 
   private _renderDropHint(dropHintIndex: number): JSX.Element {
     const classNames = this._classNames;
+    const IconComponent = this.props.useFastIcons ? FontIcon : Icon;
     return (
       <div key={'dropHintKey'} className={classNames.dropHintStyle} id={`columnDropHint_${dropHintIndex}`}>
-        <Icon
+        <IconComponent
           key={`dropHintCaretKey`}
           aria-hidden={true}
           data-is-focusable={false}
@@ -636,10 +635,8 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
       </div>
     );
   }
-  private _onRenderColumnHeaderTooltip = (
-    tooltipHostProps: ITooltipHostProps,
-    defaultRender?: IRenderFunction<ITooltipHostProps>
-  ): JSX.Element => {
+
+  private _onRenderColumnHeaderTooltip = (tooltipHostProps: ITooltipHostProps): JSX.Element => {
     return <span className={tooltipHostProps.hostClassName}>{tooltipHostProps.children}</span>;
   };
 
@@ -853,7 +850,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     }
   }
 
-  private _onToggleCollapseAll(): void {
+  private _onToggleCollapseAll = (): void => {
     const { onToggleCollapseAll } = this.props;
     const newCollapsed = !this.state.isAllCollapsed;
     this.setState({
@@ -862,7 +859,7 @@ export class DetailsHeaderBase extends BaseComponent<IDetailsHeaderBaseProps, ID
     if (onToggleCollapseAll) {
       onToggleCollapseAll(newCollapsed);
     }
-  }
+  };
 }
 
 function getLegacyColumnReorderProps(columnReorderOptions: IColumnReorderOptions): IColumnReorderHeaderProps {