+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsList.types.ts
@@ -322,6 +322,12 @@ export interface IDetailsListProps extends IBaseProps<IDetailsList>, IWithViewpo
    * Whether to animate updates
    */
   enableUpdateAnimations?: boolean;
+
+  /**
+   * Whether to use fast icon and check components. The icons can't be targeted by customization
+   * but are still customizable via class names.
+   */
+  useFastIcons?: boolean;
 }
 
 /**