+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRowCheck.styles.ts
@@ -2,6 +2,7 @@ import { getGlobalClassNames, getFocusStyle } from '../../Styling';
 import { IDetailsRowCheckStyleProps, IDetailsRowCheckStyles } from './DetailsRowCheck.types';
 import { DEFAULT_ROW_HEIGHTS } from './DetailsRow.styles';
 import { HEADER_HEIGHT } from './DetailsHeader.styles';
+import { CheckGlobalClassNames } from '../Check/Check.styles';
 
 const GlobalClassNames = {
   root: 'ms-DetailsRow-check',
@@ -10,7 +11,7 @@ const GlobalClassNames = {
 };
 
 export const getStyles = (props: IDetailsRowCheckStyleProps): IDetailsRowCheckStyles => {
-  const { theme, className, isHeader, selected, anySelected, canSelect, compact, isVisible } = props;
+  const { theme, className, isHeader, selected, anySelected, canSelect, compact, isVisible, useGlobalCheckHostClass } = props;
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
   const { rowHeight, compactRowHeight } = DEFAULT_ROW_HEIGHTS;
 
@@ -26,6 +27,7 @@ export const getStyles = (props: IDetailsRowCheckStyleProps): IDetailsRowCheckSt
       isHeader && classNames.isHeader,
       getFocusStyle(theme),
       theme.fonts.small,
+      useGlobalCheckHostClass && CheckGlobalClassNames.checkHost,
       {
         display: 'flex',
         alignItems: 'center',
@@ -36,20 +38,18 @@ export const getStyles = (props: IDetailsRowCheckStyleProps): IDetailsRowCheckSt
         background: 'none',
         backgroundColor: 'transparent',
         border: 'none',
-        opacity: 0,
+        opacity: isCheckVisible ? 1 : 0,
         height: height,
         width: 40,
         padding: 0,
-        margin: 0,
+        margin: 0
+      },
+      useGlobalCheckHostClass && {
         selectors: {
           '&:hover': {
             opacity: 1
           }
         }
-      },
-
-      isCheckVisible && {
-        opacity: 1
       }
     ],